Tests: Workaround issues with QUnit Chrome bridge
The Chrome bridge from `grunt-contrib-qunit` is now getting injected into
every single iframe, including an empty one that has no intention of
running QUnit tests. Since that bridge requires QUnit, it fails with an
error in such cases. Workaround the issue by wrapping the bridge in
another function that bails early if QUnit is not defined.